# Remove underscore regexps