--- a/go.mod
+++ b/go.mod
@@ -2,8 +2,4 @@ module github.com/pelletier/go-toml
 
 go 1.12
 
-require (
-	github.com/BurntSushi/toml v0.3.1 // indirect
-	github.com/davecgh/go-spew v1.1.1
-	gopkg.in/yaml.v2 v2.3.0 // indirect
-)
+require github.com/davecgh/go-spew v1.1.1